requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asyncpg==0.30.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.7.14
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.2
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
dnspython==2.7.0
dulwich==0.22.8
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastjsonschema==2.21.2
filelock==3.18.0
findpython==0.6.3
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.7.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
keyring==25.6.0
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pbs-installer==2025.9.2
peft==0.16.0
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.4.0
poetry==2.1.4
poetry-core==2.1.3
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.2
RapidFuzz==3.14.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
safetensors==0.5.3
sentencepiece==0.2.0
sentry-sdk==2.33.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.1
sympy==1.14.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
trove-classifiers==2025.8.26.11
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.35.0
virtualenv==20.32.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
zstandard==0.24.0
